migrations and initial data must work with an empty database
Some of the migrations assumed that the live site data was
present. The initial data is still somewhat incomplete; the answer
groups that I added do not have any answers, but I'd like to find out
how the initial data set was originally created before I do more
manual work on it.