use django-compressor to minify our js and css
django-compressor is pretty clever, and lets us leave the js/css in
the html, rather than having to move it to files first. A simple
modification to the base template catches most of the js and css we
use. Some refinement will be needed, but this is already pretty good.