preprocess the q&a data, in order to simplify the template
The database gives us a flat list of questions and answers, we build
nested set of dicts to organize them; this eliminates the need for the
template to contain a state machine, making it readable and logical.

This change breaks the 'hide unanswered questions' option, since that
has the side effect of not including any questions in the list,
breaking the code that builds the data structure.

The existing code already did a number of preprocessing steps to
gather other data; these ought to be folded in and the data added to
the data structure instead of to parrallel arrays. Once this is done
the preprocessing step can be moved to get_ahj_data and a fair amount
of duplicate code removed.